rhi: Fix up the instancing step rate type
This is a UINT/uint32_t/GLuint/NSUInteger in all APIs (with Metal
being special due to being 64-bit in 64-bit apps whereas all others
are 32-bit always, at least on 64-bit Windows)

As the stride is already an uint32, follow suit for the step rate.
There was no reason to have this as int in the first place.

As an added bonus, some casts, that were previously needed when
mapping to the underlying API reality, can now be removed.

Change-Id: I8e0eef037bd795b637578dfc3e59dc2efaa5976c
Reviewed-by: Christian Strømme <christian.stromme@qt.io>