QDataBuffer: statically assert !QTypeInfo<value_type>::isComplex
The implementation assumes this, so assert it.

I've tried the moral equivalent of is_pod first, but it turns out most
of the concrete value_types aren't trivial, so let's use our own trait
for now.

Put the static_assert into the dtor to give users a fighting chance to
Q_DECLARE_TYPEINFO before first instantiation (the dtor can usually be
declared out-of-line to delay instantiation of the assertion).

Change-Id: I6c03f7c554b9e07608dee639ab0abc2ddfe3c28a
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Knoll <lars@knoll.priv.no>